+++ packages/plugin-link-preview/web/src/getToolbarButtons.tsx
@@ -5,6 +5,7 @@ import { PLUGIN_TOOLBAR_BUTTON_ID } from 'wix-rich-content-editor-common';
 import LinkPreviewSettingsModal from './modals/SettingsModal';
 import { LinkPreviewSettingsButton } from './toolbar/SettingsButton';
 import { RemovePreviewButton } from './toolbar/RemovePreviewButton';
+import { TIPTAP_EMBED_TYPE, TIPTAP_LINK_PREVIEW_TYPE } from 'wix-rich-content-common';
 
 const selectedNodeResolver = {
   id: 'selectedNode',
@@ -17,6 +18,17 @@ const selectedNodeResolver = {
   },
 };
 
+const isLinkPreviewSelectedResolver = {
+  id: 'isLinkPreviewSelected',
+  resolve: content => {
+    if (Array.isArray(content) && content.length > 0) {
+      return content[0].type?.name === TIPTAP_LINK_PREVIEW_TYPE;
+    } else {
+      return false;
+    }
+  },
+};
+
 export const getToolbarButtons = (config, services): ToolbarButton[] => {
   const { modals } = services;
   return [
@@ -24,10 +36,11 @@ export const getToolbarButtons = (config, services): ToolbarButton[] => {
       id: 'removePreview',
       dataHook: 'baseToolbarButton_replaceToLink',
       command: ({ editorCommands, node }) => {
+        const link = node.type.name === TIPTAP_EMBED_TYPE ? node.attrs.src : node.attrs.link.url;
         editorCommands
           .chain()
           .focus()
-          .insertContent(node.attrs.link.url + ' ')
+          .insertContent(link + ' ')
           .run();
       },
       attributes: {
@@ -57,6 +70,7 @@ export const getToolbarButtons = (config, services): ToolbarButton[] => {
       },
       attributes: {
         selectedNode: selectedNodeResolver,
+        visible: isLinkPreviewSelectedResolver,
       },
       renderer: toolbarItem => <LinkPreviewSettingsButton toolbarItem={toolbarItem} />,
     },