+++ CHANGELOG.md
@@ -23,6 +23,8 @@
   - [#4348](https://github.com/wix/ricos/pull/4348) debug-mode prop added to RicosEditor (tiptap)
 - `gallery`
   - [#4346](https://github.com/wix/ricos/pull/4346) changling layout doesnt remove focus (tiptap)
+- `link-preview`
+  - [#4347](https://github.com/wix/ricos/pull/4347) social-embed toolbar buttons
 
 </details>